Make agent log capture test buffers concurrency-safe
