+++ internal/services/agent/log_buffer_test.go
@@ -1,0 +1,31 @@
+package agent_test
+
+import (
+	"bytes"
+	"sync"
+)
+
+// syncBuffer is a thread-safe bytes.Buffer wrapper for capturing zerolog output
+// from code paths that may log from background goroutines.
+type syncBuffer struct {
+	mu  sync.Mutex
+	buf bytes.Buffer
+}
+
+func (s *syncBuffer) Write(p []byte) (int, error) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	return s.buf.Write(p)
+}
+
+func (s *syncBuffer) String() string {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	return s.buf.String()
+}
+
+func (s *syncBuffer) Bytes() []byte {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	return append([]byte(nil), s.buf.Bytes()...)
+}
